+++ src/Strategy/Leiningen.hs
@@ -241,12 +241,13 @@ ednVecToMap :: EDN.EDNVec -> Parser EDN.EDNMap
 ednVecToMap = go Map.empty
   where
     go :: EDN.EDNMap -> EDN.EDNVec -> Parser EDN.EDNMap
+    -- TODO: refactor this to not use match guards
     go m vec
       | V.null vec = pure m
       | otherwise = do
-        key <- EDN.vecGet 0 vec
-        value <- EDN.vecGet 1 vec
-        go (Map.insert key value m) (V.drop 2 vec)
+          key <- EDN.vecGet 0 vec
+          value <- EDN.vecGet 1 vec
+          go (Map.insert key value m) (V.drop 2 vec)
 
 -- | The FromEDN type for lein deps output
 newtype Deps = Deps