+++ test/Go/TransitiveSpec.hs
@@ -1,14 +1,117 @@
 module Go.TransitiveSpec (spec) where
 
 import Control.Algebra (run)
-import DepTypes
-import GraphUtil
-import Strategy.Go.Transitive
+import DepTypes (DepType (GoType), Dependency (..), VerConstraint (CEq))
+import GraphUtil (expectDeps, expectEdges)
+import Strategy.Go.Transitive as Transitive (
+  GoListPkgImportPath (GoListPkgImportPath),
+  GoPackageReplacement (..),
+  Module (Module, modPath, modReplacement, modVersion),
+  NormalizedImportPath (NormalizedImportPath),
+  Package (..),
+  graphTransitive,
+  normalizeImportPaths,
+ )
+
+import Data.Coerce (coerce)
 import Strategy.Go.Types (graphingGolang)
-import Test.Hspec
+import Test.Hspec (Spec, describe, it, shouldBe)
 
 spec :: Spec
-spec = spec_packageToModule
+spec = do
+  spec_packageToModule
+  performsPackageReplacementSpec
+
+performsPackageReplacementSpec :: Spec
+performsPackageReplacementSpec =
+  describe "Package list module replacements" $ do
+    it "replaces module name and version when the replacement isn't in an import list" $ do
+      let normalized = normalizeImportPaths unimportedPackageReplacement
+          graph = run $ graphingGolang (graphTransitive normalized)
+      expectDeps [replacedModuleDep, nonReplacedPackageDep] graph
+      expectEdges [(replacedModuleDep, nonReplacedPackageDep)] graph
+
+    it "replaces module name and version when the replacement is in an import list" $ do
+      let normalized = normalizeImportPaths importedPackageReplacement
+          graph = run $ graphingGolang (graphTransitive normalized)
+      expectDeps [replacedModuleDep, nonReplacedPackageDep] graph
+      expectEdges [(nonReplacedPackageDep, replacedModuleDep)] graph
+
+replacedModuleDep :: Dependency
+replacedModuleDep =
+  Dependency
+    { dependencyType = GoType
+    , dependencyName = "github.com/example/bar"
+    , dependencyVersion = Just $ CEq "2.0.0"
+    , dependencyLocations = []
+    , dependencyEnvironments = mempty
+    , dependencyTags = mempty
+    }
+
+nonReplacedPackageDep :: Dependency
+nonReplacedPackageDep =
+  Dependency
+    { dependencyType = GoType
+    , dependencyName = "github.com/example/baz"
+    , dependencyVersion = Nothing
+    , dependencyLocations = []
+    , dependencyEnvironments = mempty
+    , dependencyTags = mempty
+    }
+
+unimportedPackageReplacement :: [Package GoListPkgImportPath]
+unimportedPackageReplacement =
+  [ Package
+      { packageImportPath = "github.com/example/foo/inner-package"
+      , packageModule =
+          Just
+            Module
+              { modPath = "github.com/example/foo"
+              , modVersion = Just "1.0.0"
+              , modReplacement =
+                  Just
+                    GoPackageReplacement
+                      { Transitive.pathReplacement = "github.com/example/bar"
+                      , Transitive.versionReplacement = "2.0.0"
+                      }
+              }
+      , packageImports = Just ["github.com/example/baz"]
+      , packageSystem = Nothing
+      }
+  , Package
+      { packageImportPath = "github.com/example/baz"
+      , packageModule = Nothing
+      , packageImports = Nothing
+      , packageSystem = Nothing
+      }
+  ]
+
+importedPackageReplacement :: [Package GoListPkgImportPath]
+importedPackageReplacement =
+  [ Package
+      { packageImportPath = "github.com/example/foo/inner-package"
+      , packageModule =
+          Just
+            Module
+              { modPath = "github.com/example/foo"
+              , modVersion = Just "1.0.0"
+              , modReplacement =
+                  Just
+                    GoPackageReplacement
+                      { Transitive.pathReplacement = "github.com/example/bar"
+                      , Transitive.versionReplacement = "2.0.0"
+                      }
+              }
+      , packageImports = Nothing
+      , packageSystem = Nothing
+      }
+  , Package
+      { packageImportPath = "github.com/example/baz"
+      , packageModule = Nothing
+      , packageImports = Just ["github.com/example/foo/inner-package"]
+      , packageSystem = Nothing
+      }
+  ]
 
 -- HACK(fossas/team-analysis#514) see Strategy.Go.Transitive for more details
 --
@@ -16,20 +119,20 @@ spec = spec_packageToModule
 -- aren't dropped from the dependency graph
 spec_packageToModule :: Spec
 spec_packageToModule =
-  describe "normalizeImportsToModules" $ do
+  describe "normalizeImportPaths" $ do
     it "should map package imports to their modules" $ do
-      let result = normalizeImportsToModules testPackages
+      let result = normalizeImportPaths testPackages
       result `shouldBe` normalizedPackages
 
     it "should prevent packages from appearing in the final graph" $ do
-      let graph = run $ graphingGolang (graphTransitive (normalizeImportsToModules testPackages))
+      let graph = run $ graphingGolang (graphTransitive (normalizeImportPaths testPackages))
       expectDeps [fooDep, barDep, bazDep] graph
       expectEdges [(fooDep, barDep), (barDep, bazDep)] graph
 
     it "should be a no-op for non-module packages" $ do
-      normalizeImportsToModules nonModulePackages `shouldBe` nonModulePackages
+      normalizeImportPaths nonModulePackages `shouldBe` coerce nonModulePackages
 
-nonModulePackages :: [Package]
+nonModulePackages :: [Package GoListPkgImportPath]
 nonModulePackages =
   [ Package
       { packageImportPath = "github.com/example/foo"
@@ -48,7 +151,7 @@ nonModulePackages =
       }
   ]
 
-testPackages :: [Package]
+testPackages :: [Package GoListPkgImportPath]
 testPackages =
   [ Package
       { packageImportPath = "github.com/example/foo"
@@ -66,6 +169,7 @@ testPackages =
             Module
               { modPath = "github.com/example/bar"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports =
           Just
@@ -81,6 +185,7 @@ testPackages =
             Module
               { modPath = "github.com/example/baz"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports = Nothing
       , packageSystem = Nothing
@@ -92,13 +197,14 @@ testPackages =
             Module
               { modPath = "github.com/example/baz"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports = Nothing
       , packageSystem = Nothing
       }
   ]
 
-normalizedPackages :: [Package]
+normalizedPackages :: [Package NormalizedImportPath]
 normalizedPackages =
   [ Package
       { packageImportPath = "github.com/example/foo"
@@ -110,12 +216,13 @@ normalizedPackages =
       , packageSystem = Nothing
       }
   , Package
-      { packageImportPath = "github.com/example/bar/some/package"
+      { packageImportPath = "github.com/example/bar"
       , packageModule =
           Just
             Module
               { modPath = "github.com/example/bar"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports =
           Just
@@ -125,12 +232,13 @@ normalizedPackages =
       , packageSystem = Nothing
       }
   , Package
-      { packageImportPath = "github.com/example/baz/other"
+      { packageImportPath = "github.com/example/baz"
       , packageModule =
           Just
             Module
               { modPath = "github.com/example/baz"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports = Nothing
       , packageSystem = Nothing
@@ -142,6 +250,7 @@ normalizedPackages =
             Module
               { modPath = "github.com/example/baz"
               , modVersion = Nothing
+              , modReplacement = Nothing
               }
       , packageImports = Nothing
       , packageSystem = Nothing